Pipeline established up to stage transition Not included yet: -Moving player to correct spot -Anything to do with sacrifice rooms
