+++ scripts/enums/rooms.lua
@@ -1,4 +1,6 @@
-return {
+local mod = GreedSpecialRooms
+
+mod.enum.rooms = {
 	[RoomType.ROOM_ARCADE] = {
         {id = 0, weight = 1},
         {id = 1, weight = 1},
