+++ scripts/libs/the-everything-function-rev1.lua
